requirements.txt
# Lists the version of Python dependencies to be installed
Flask==2.2.3
mysql-connector-python==8.0.31